backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic-settings
python-multipart
googlemaps
numpy
torch
opencv-python
facenet-pytorch
requests
geopy
